Cleanup engine precompile
